chore: refactor and clean up
