+++ packages/logos/src/types/index.ts
@@ -3,11 +3,3 @@ export interface CarLogo {
   url: string;
   filename: string;
 }
-
-export interface LogoMetadata {
-  brand: string;
-  filename: string;
-  url: string;
-  createdAt: string;
-  fileSize?: number;
-}
