+++ packages/logos/src/utils/normalise-brand-name.ts
